// src/app/model/kpi.ts
export interface kpi {
  Title: string;
  Modified: string;	
  Created: string;
  Department: string;
  Relevance: string;
  Code: string;
  CoreValues: string;	
  Levels: string;
  Category: string;
  KPI: string;
  Impact: string;
  Plan: string;
  Weightage: string;
  Metric: string;
  CreatedBy: string;
  ModifiedBy: string;
  status:string;
  RatingDefinationTwo:string;
  RatingDefinationThree:string;
  RatingDefinationFour:string;
  Time:string;
  MidYearRating:string;
  AnnualYearRating:string;
  }